+++ internal/amf/ngap/message/forward_ie.go
@@ -18,7 +18,7 @@ import (
 )
 
 func AppendPDUSessionResourceSetupListSUReq(list *ngapType.PDUSessionResourceSetupListSUReq,
-	pduSessionID int32, snssai *models.Snssai, nasPDU []byte, transfer []byte,
+	pduSessionID uint8, snssai *models.Snssai, nasPDU []byte, transfer []byte,
 ) {
 	var item ngapType.PDUSessionResourceSetupItemSUReq
 	item.PDUSessionID.Value = int64(pduSessionID)
@@ -36,7 +36,7 @@ func AppendPDUSessionResourceSetupListSUReq(list *ngapType.PDUSessionResourceSet
 	list.List = append(list.List, item)
 }
 
-func AppendPDUSessionResourceSetupListHOReq(list *ngapType.PDUSessionResourceSetupListHOReq, pduSessionID int32, snssai *models.Snssai, transfer []byte) {
+func AppendPDUSessionResourceSetupListHOReq(list *ngapType.PDUSessionResourceSetupListHOReq, pduSessionID uint8, snssai *models.Snssai, transfer []byte) {
 	var item ngapType.PDUSessionResourceSetupItemHOReq
 	item.PDUSessionID.Value = int64(pduSessionID)
 	snssaiNgap, err := util.SNssaiToNgap(snssai)
@@ -49,7 +49,7 @@ func AppendPDUSessionResourceSetupListHOReq(list *ngapType.PDUSessionResourceSet
 	list.List = append(list.List, item)
 }
 
-func AppendPDUSessionResourceSetupListCxtReq(list *ngapType.PDUSessionResourceSetupListCxtReq, pduSessionID int32, snssai *models.Snssai, nasPDU []byte, transfer []byte) {
+func AppendPDUSessionResourceSetupListCxtReq(list *ngapType.PDUSessionResourceSetupListCxtReq, pduSessionID uint8, snssai *models.Snssai, nasPDU []byte, transfer []byte) {
 	var item ngapType.PDUSessionResourceSetupItemCxtReq
 	item.PDUSessionID.Value = int64(pduSessionID)
 	snssaiNgap, err := util.SNssaiToNgap(snssai)
@@ -67,7 +67,7 @@ func AppendPDUSessionResourceSetupListCxtReq(list *ngapType.PDUSessionResourceSe
 }
 
 func AppendPDUSessionResourceModifyListModReq(list *ngapType.PDUSessionResourceModifyListModReq,
-	pduSessionID int32, nasPDU []byte, transfer []byte,
+	pduSessionID uint8, nasPDU []byte, transfer []byte,
 ) {
 	var item ngapType.PDUSessionResourceModifyItemModReq
 	item.PDUSessionID.Value = int64(pduSessionID)
@@ -89,7 +89,7 @@ func AppendPDUSessionResourceModifyListModCfm(list *ngapType.PDUSessionResourceM
 }
 
 func AppendPDUSessionResourceToReleaseListRelCmd(list *ngapType.PDUSessionResourceToReleaseListRelCmd,
-	pduSessionID int32, transfer []byte,
+	pduSessionID uint8, transfer []byte,
 ) {
 	var item ngapType.PDUSessionResourceToReleaseItemRelCmd
 	item.PDUSessionID.Value = int64(pduSessionID)
